Fix division in result2 for accurate calculations
Updated the calculation of `result2` in the `Main`

+++ C#2/Program.cs
@@ -5,7 +5,7 @@
         static void Main(string[] args)
         {
             double result1 = (3 + 5 * (6 + 3) - 8 * 3 + 1) / (21 + 50 /(double) (3 + 4 * (1 + 2)));
-            double result2 = ((0.25 - 0.12) * 0.81 + 0.132 - 2.7 / 3.1) / (3 / 7 + 2.97 * (8.05 - 8.1 * 6.07));
+            double result2 = ((0.25 - 0.12) * 0.81 + 0.132 - 2.7 / 3.1) / ((double)3 / 7 + 2.97 * (8.05 - 8.1 * 6.07));
             Console.WriteLine("Результаты вычислений:");
             Console.WriteLine(result1);
             Console.WriteLine(result2);
